Reorganize back-end
