+++ packages/back/src/controllers/auth.controller.ts
@@ -1,23 +1,22 @@
 import { Request, Response } from 'express'
 
-import { UserController } from '../database'
+import { checkUserLogin, registerUser } from '../db'
 import { runRequestValidator } from '../utils'
 import type { User } from '../types'
 
-export class AuthController {
-  public async login(req: Request<{}, any, Pick<User, 'email' | 'password'>>, res: Response) {
-    runRequestValidator(req)
+export const loginRequestHandler = async (req: Request<{}, any, Pick<User, 'email' | 'password'>>, res: Response) => {
+  runRequestValidator(req)
 
-    const { email, password } = req.body
-    res.json({ data: await UserController.login(email, password) })
-  }
-
-  public async register(req: Request<{}, any, Pick<User, 'email' | 'name' | 'password'>>, res: Response) {
-    runRequestValidator(req)
-
-    const { email, password, name } = req.body
-    res.json({ data: await UserController.register(email, password, name) })
-  }
+  const { email, password } = req.body
+  res.json({ data: await checkUserLogin(email, password) })
 }
 
-export default new AuthController()
+export const registerRequestHandler = async (
+  req: Request<{}, any, Pick<User, 'email' | 'name' | 'password'>>,
+  res: Response
+) => {
+  runRequestValidator(req)
+
+  const { email, password, name } = req.body
+  res.json({ data: await registerUser(email, password, name) })
+}
