+++ packages/back/src/controllers/profile.controller.ts
@@ -1,24 +1,20 @@
 import { Response } from 'express'
 
-import { UserController } from '../database'
+import { editUser, deleteUser } from '../db'
 import type { RequestAuthed, User } from '../types'
 
-export class ProfileController {
-  public async getProfile(req: RequestAuthed, res: Response) {
-    const data = req.userDoc
-    delete data.uploads
-    res.json({ data })
-  }
-
-  public async editProfile(req: RequestAuthed<{}, any, Pick<User, 'name'>>, res: Response) {
-    const newProfile = await UserController.edit(req.session._id, req.body)
-    res.json({ data: newProfile })
-  }
-
-  public async deleteAccount(req: RequestAuthed, res: Response) {
-    const { _id } = await UserController.delete(req.session._id)
-    res.json({ data: { _id } })
-  }
+export const getProfileRequestHandler = async (req: RequestAuthed, res: Response) => {
+  const data = req.userDoc
+  delete data.uploads
+  res.json({ data })
 }
 
-export default new ProfileController()
+export const editProfileRequestHandler = async (req: RequestAuthed<{}, any, Pick<User, 'name'>>, res: Response) => {
+  const newProfile = await editUser(req.session._id, req.body)
+  res.json({ data: newProfile })
+}
+
+export const deleteAccountRequestHandler = async (req: RequestAuthed, res: Response) => {
+  const { _id } = await deleteUser(req.session._id)
+  res.json({ data: { _id } })
+}
