+++ packages/back/src/index.ts
@@ -1,7 +1,3 @@
 import { initServer } from './server'
 
-const setup = async () => {
-  await initServer()
-}
-
-setup()
+initServer()
